First main code
A Code that scrapes from flashscore site all matches and turn it into a pandas dataframe. Using beautifulsoup, selenium chrome webdriver and of course, pandas.

There is a function "Nextday" that clicks in the "tomorrow" button, so the code can swith days to get the matches from different days.

In this version, the code only build a dataframe with the matches like it is being appearing in flashscore website. 

For next versions I intent to filter matches of my interest, connect with an API to google calendar and to automate to run it everyday. Although this is just a hobby project.